Use PackageManager.setComponentEnabledSetting() instead of requiring root. Use WRITE_SECURE_SETTINGS when available to auto-manage the "ADB Debugging" settings.
Many users are confused about the requirement to enable ADB debugging.
The initial message instructing the user to enable it doesn't seem to be
enough. It's also a problem in that it can be a security liablity. I
have a variety of ideas to improve this situation, and this is the first
change in that regard.

If we're signed with a system certificate, we can use
setComponentEnabledState() rather than going the root-route. If we're
installed on the system partition, we can auto-enable the ADB Debugging
setting.