Always use a timeout.
While we shouldn't really need the timeout anymore with ADB being
enabled by us automatically, there are still those rarse cases where we
see problems (as described in e9e7a203e97a24d7b8b4077a903542455fa19986,
for example). Using a timeout in all cases ensures that the user
experience in case there are problems isn't impacted too much.

It might be helpful though to indicate to the user somehow that a
timeout is employed, and that potentially it should be investigated why
that happens.