Added GUI to toggle between the two grouping options.
The icon is ic_menu_windows.png from the Browser app. Another option
would have been a custom icon based on the Browser's ic_menu_list, with
reduced width for the lower bars (indicating a child item), and 180°
rotation, so that the darkest item is on the top. It would be similar to
the "Sort A-Z" icon in the system "Manage Applications" activity, but
with only two different item lengths.