Use the Fragment API to show the receiver details dialog.
Why? The terrible hack we used to change the Enabled/Disabled label in
onPrepare() stopped working on 4.0. Rather than tracking the problem
down and finding another workaround, take steps in the right direction.

A lot of ugly code is now much nicer; a lot of other code is still ugly.
The subsequent dialogs that might occur are still using the showDialog()
approach, so now we have the Fragment heavily calling into the Activity.

Ideally, in the future this would be further redesign, with the explicit
goal of getting rid of the mLastSelectedReceiver status attributes.

The compatibility package is used to provide 1.6 support. This commit
drops compatibility with 1.5, with the compatibility package does not
support.

As part of this, a new "compat" package was created, and the
FixedExpandableListView class moved there as well.