Remove toggle warnings, allows some refactoring.
The old dialog system is now deprecated, so I wanted to get rid of it. Also,
it always sucked. However, showing three different types of alert dialogs
via fragments, and passing along the selected receiver information, maintaining
it through screen rotation etc. did not sound fun either.

Instead, I opted to gt rid of the system app warning dialog, instead integrating
it into the details dialog, and got rid of the Google Talk warning for good
- I'm not even sure it's still applicable, but if so, I'd like to bring it
back as a more general, community-edited system of undesirable effects.

This all allows us to get rid of the ugly remembering the "LastSelectedEvent"
stuff.

One dialog is left that needs to be ported.